Extract helpers and profile code from TypeScript writer
Split writer.ts (1651 lines) into three files:
- utils.ts: standalone utility functions and constants
- profile.ts: profile-specific helpers and extracted class methods
- writer.ts: TypeScript class with non-profile methods (317 lines)
